+++ Shared/DevicesLib/Entities/Component/Cpu/ICpu.cs
@@ -4,9 +4,9 @@ namespace DevicesLib.Entities.Component.Cpu;
 
 public interface ICpu : IComponent
 {
-    public float OneMinuteLoad { get; set; }
-    public float FiveMinuteLoad { get; set; }
-    public float FifteenMinuteLoad { get; set; }
+    public int OneMinuteLoad { get; set; }
+    public int FiveMinuteLoad { get; set; }
+    public int FifteenMinuteLoad { get; set; }
     
     public List<ICpuCore> Cores { get; set; }
 }
